feat(airplay_receiver): port AirPlay2 HomeKit pairing + ChaCha20-Poly1305 crypto

+++ components/airplay_receiver/airplay_receiver.cpp
@@ -7,8 +7,8 @@ namespace airplay_receiver {
 static const char *const TAG = "airplay_receiver";
 
 void AirPlayReceiver::setup() {
-  ESP_LOGCONFIG(TAG, "AirPlayReceiver '%s' buffer_size=%u", this->airplay_name_.c_str(),
-                this->buffer_size_);
+  ESP_LOGCONFIG(TAG, "AirPlayReceiver '%s' buffer_size=%lu", this->airplay_name_.c_str(),
+                (unsigned long) this->buffer_size_);
 }
 
 void AirPlayReceiver::loop() {
@@ -17,7 +17,7 @@ void AirPlayReceiver::loop() {
 
 void AirPlayReceiver::dump_config() {
   ESP_LOGCONFIG(TAG, "AirPlayReceiver name='%s'", this->airplay_name_.c_str());
-  ESP_LOGCONFIG(TAG, "  buffer_size=%u", this->buffer_size_);
+  ESP_LOGCONFIG(TAG, "  buffer_size=%lu", (unsigned long) this->buffer_size_);
 }
 
 }  // namespace airplay_receiver
